requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-magic==0.4.27
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
google-cloud-vision==3.5.0
google-auth==2.25.0